Cosine

+++ CosinusTest/cosinus.c
@@ -3,14 +3,14 @@
 
 #include "cosinus.h"
 
-long int silnia(long int n){
+unsigned long int silnia(long int n){
   if (n == 0) return 1;
   else return n*silnia(n-1);
 }
 
-double potega(double x, long int n){
+long double potega(long double x, long int n){
 
-  double wynik = 1.0;
+  long double wynik = 1.0;
 
   for(long int i=1; i<=n; i++){
       wynik*=x;
@@ -19,24 +19,25 @@ double potega(double x, long int n){
 }
 
 //obliczanie z rozwinięcia Maclaurina od przodu
-double cosMaclaurinFront(double x, long int k){
+long double cosMaclaurinFront(long double x, long int k){
 
-  double sum = 0.0;
+  long double sum = 0.0;
 
   for(long int i=0; i<=k; i++){
       sum += potega(-1, i) * potega(x, 2*i) / silnia(2*i);
   }
+
   return sum;
 }
 
 //obliczanie z rozwinięcia Maclaurina od tyłu
-double cosMaclaurinBack(double x, long int k){
+long double cosMaclaurinBack(long double x, long int k){
 
-  double sum = 0.0;
-  double tmp[k];
+  long double sum = 0.0;
+  long double tmp[k];
 
   for(long int i=0; i<=k; i++){
-      tmp[i] = potega(-1, i) * potega(x, 2*i) / silnia(2*i);
+      tmp[i] = potega(-1, i) * potega(x, 2.0*i) / silnia(2.0*i);
   }
 
   for(long int i=k; i>=0; i--){
@@ -47,13 +48,13 @@ double cosMaclaurinBack(double x, long int k){
 }
 
 //obliczanie na podstawie poprzedniego wyrazu od przodu
-double cosFront(double x, long int k){
+long double cosFront(long double x, long int k){
 
-  double t   = 1.0;
-  double sum = 1.0;
+  long double t   = 1.0;
+  long double sum = 1.0;
 
   for(long int i=1; i<=k; i++){
-    t = t * (-1)*x*x / (2*i*(2*i-1));
+    t = t * (-1)*x*x / (2.0*i*(2.0*i-1.0));
     sum = sum + t;
   }
 
@@ -61,15 +62,15 @@ double cosFront(double x, long int k){
 }
 
 //obliczanie na podstawie poprzedniego wyrazu od tyłu
-double cosBack(double x, long int k){
+long double cosBack(long double x, long int k){
 
-  double sum = 0.0;
-  double tmp[k+1];
+  long double sum = 0.0;
+  long double tmp[k+1];
 
   tmp[0] = 1.0;
 
   for(long int i=1; i<=k; i++){
-    tmp[i] = tmp[i-1] * (-1)*x*x / (2*i*(2*i-1));
+    tmp[i] = tmp[i-1] * (-1)*x*x / (2.0*i*(2.0*i-1.0));
   }
 
   for(long int i=k; i>=0; i--){
